Added Table View to the the View Controller and configured its position.
Updated the ViewController class declaration to include the UITableViewDelegate, UITableViewDataSource references which are required providing a data source and delegate for TableView.

Added an @IBOutlet named tableView for the UITableView.

Set the dataSource to value of self for the tableView.

Set the delegate to value of self for the tableView.

Removed the didReceiveMemoryWarning() function.

Added the numberOfRowsInSection function to return the required number of row for the tableView and set to return 10.

Added cellForRowAt indexPath which is required for setting the value of each cell in tableView.

Add a constant named cell and set to equal value of UITableViewCell () class.

Set the text value for the labelText of the cell to “Hello”

Returned the value of cell.