Added a new swift file named CompleteTaskViewController.
Added didSelectRowAt indexPath to preform a segue when a task is tapped.

Updated the prepare(for segue:  function with an if statement to handle multiple segue.

Created an @IBOutlet for the taskLable.

Created a property variable named task of type Task () class.

Set the taskLable.text to equal the of value task.name.

Created an @IBAction for completeTapped function.