*** App is working but no CoreData :( ***
Added a new variable named selectedIndex to capture the current tapped task index.

Updated the didSelectRowAt indexPath to set the value of the selectedIndex to the value of indexPath.row.

Created a property variable named previousVC and set it to equal to TasksViewController ().

Updated the viewDidLoad() function with an if statement to add an “!” mark to the name of the the task if it’s important.

Updated the completeTapped function to reload the Table View of the TasksViewController after removing the tapped task and to move back instantly to the TasksViewController using the navigationController!.popViewController(animated: true) method.